Speed up integral images

+++ mpicbg/src/main/java/mpicbg/ij/integral/NormalizeLocalContrast.java
@@ -85,7 +85,7 @@ public class NormalizeLocalContrast extends BlockStatistics
 			{
 				final int xMin = Math.max( -1, x - blockRadiusX - 1 );
 				final int xMax = Math.min( w, x + blockRadiusX );
-				final long bs = ( xMax - xMin ) * bh;
+				final double bs = ( xMax - xMin ) * bh;
 				final double scale = 1.0 / bs;
 				final double sum = sums.getDoubleSum( xMin, yMin, xMax, yMax );
 				final int i = row + x;
@@ -130,7 +130,7 @@ public class NormalizeLocalContrast extends BlockStatistics
 			{
 				final int xMin = Math.max( -1, x - blockRadiusX - 1 );
 				final int xMax = Math.min( w, x + blockRadiusX );
-				final long bs = ( xMax - xMin ) * bh;
+				final double bs = ( xMax - xMin ) * bh;
 				final double scale1 = 1.0 / ( bs - 1 );
 				final double scale2 = 1.0 / ( bs * bs - bs );
 				final double sum = sums.getDoubleSum( xMin, yMin, xMax, yMax );
@@ -178,7 +178,7 @@ public class NormalizeLocalContrast extends BlockStatistics
 			{
 				final int xMin = Math.max( -1, x - blockRadiusX - 1 );
 				final int xMax = Math.min( w, x + blockRadiusX );
-				final long bs = ( xMax - xMin ) * bh;
+				final double bs = ( xMax - xMin ) * bh;
 				final double scale = 1.0 / bs;
 				final double scale1 = 1.0 / ( bs - 1 );
 				final double scale2 = 1.0 / ( bs * bs - bs );
